refactor(app/main/plugins): move `plugins` test
