+++ app/main/plugins/plugins.spec.js
@@ -1,9 +1,10 @@
+jest.dontMock('electron');
+
 import path from 'path';
 import { utils } from 'dext-core-utils';
-import { CORE_PLUGIN_PATH } from '../../../app/constants';
-import plugins from '../../../app/main/plugins';
+import { CORE_PLUGIN_PATH } from '../../constants';
+import plugins from './plugins';
 
-jest.dontMock('electron');
 jest.mock('plist');
 jest.mock('conf');
 jest.mock('fs');
